deviceinstall: split example frontend into modules
